Problem.pm now works (kindof) to serve PG problems.  There are configuration variables at the top that have to be customized to an individuals directory. (They could be placed in a config file to avoid having to update them constantly, but soon these values should be set by information from the course and problem environments.)
The code is still rather delicate.  It seems to work with some problems and not
with others.  I suspect that the macro files are not being properly read in
and that perhaps the warning mechanisms are not yet working properly.

Also beware the line ending problem.  Some of these files were created on a
mac (using unix).  I've had completely  mysterious errors that were fixed
by checking the line endings.

There is also a lot of code, particulary the error reporting code which should be put
in its own module.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@392 c0722133-6baf-4dd8-8699-98d999cd4f06