Changed warning messages when dvipng doesn't work from die to warn. This gives more infformation back to the webbrowser and still prints the same information to the log.
Removed the workaround for dvipng's problems with labeling its dvi
output.  This problem was fixed in the latest version of dvipng which is
now being used on webwork-db and webwork.

There is still a problem that dvipng is being used to display answers at
all times (even if the mode switch is set to tth mode )


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@624 c0722133-6baf-4dd8-8699-98d999cd4f06