Considerable progress has been made.  The select definition button and choose problem  buttons work now, although the code needs to be de-crufted. For some reason Safari will not recognize \r\n (newlines) in the textarea although iCab does.  What is the accepted way to enter text with newlines into the data for a textarea?
The view problem button doesn't work yet.  i want this to come up in a new window.
Must I use javaScript for this or is there another way to force a new target window
for the result of pusing the view problem button?  How can one redirect to a particular
problem?

There is still considerable debugging code in this file.

--Mike

CVe: ----------------------------------------------------------------------


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@868 c0722133-6baf-4dd8-8699-98d999cd4f06