The select set and the choose problem button now work fairly well. There is a static link allowing you to view the last chosen problem via the problem library at webhost.  This works for now, but will break if the class template file is out of sync with the template file used by the "problem library" program at webhost.
The save button, which involves checking the syntax of the problem list has not been implemented. (There is code in profBuildProbSets to help with this.)

The view entire problem set (in pdf form) is not been implemented.
It should be possible to redirect this to the hardcopy feature somehow?
or to the PDF feature of the "problem library".

Other aspects of the user inface are still problematic.  We'll need to
play with it for awhile to see which behaviors must be changed.

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@869 c0722133-6baf-4dd8-8699-98d999cd4f06