Editor now  fills a blank textarea from the named problem set if it exists.
There are a number of user interface issues to be tackled here.
The behavior is: if the textarea is empty and a setName is defined
then the textarea is filled from the setNAme.def file (eventually
it will be filled from the database )  This behavior will probably
not be transparent to the user.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@884 c0722133-6baf-4dd8-8699-98d999cd4f06