Problem Editor now creates a temporary file to edit ( adding .tmp to the file name) Refresh causes a redirect to the Problem.pm with a GET command. Some of the arguments are unnecessary and need to be cleaned up.
The seed and display mode in the problem editor are not properly
initialized from the problem set.

Problem.pm has been modified to check for these arguments and to read
the temporary source file, seed and displayMode.  These are passed
through to a PG.pm object.

PG also needed to be modified so that the source_file and seed values
are overridden.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@926 c0722133-6baf-4dd8-8699-98d999cd4f06