Added a new template hook "options" which allows one to reset the display options.
Modified viewOptions to conform to a "vertical" look for the radio buttons in the redisplay options.
Modified some of the labels for the same reason.
The horizontal spacing is a bit of kludge (I can't guarantee no line breaks for each of the radio button
labels).  Using underbars instead of spaces in the labels would fix this, but...

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@940 c0722133-6baf-4dd8-8699-98d999cd4f06