The problem set editor now works for 0 users (global only) or 1 user. Multi-user functionality is coming over the weekend.
To cause the editor to give you the ability to edit a student's record,
add editForStudent=name (where name is the student's login) to the
parameters sent to the ProblemSetEditor content generator.

Note that if you are using the GDBM backend, editing the "default"
professor records is the same as editing the global values, which can
surprise you if you're not ready for it.
-Dennis


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@959 c0722133-6baf-4dd8-8699-98d999cd4f06