Made small modifications to allow Problem.pm to read a problem that has not yet been assigned.
 TO DO (and I'll need help with this)
Convert a global set to a global user set (and the global problem to a
global user problem).  I need to know how to specify the appropriate
classes.

Once we are guaranteed a record of the type returned by getGlobalUser
....

stuff missing defaults (e.g. psvn for set, and num_correct (answers) for
problem) in the newly created global_user record objects.(this part I
can do.)

At that point the rest of Problem.pm should be able to proceed as before
without any changes.

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@993 c0722133-6baf-4dd8-8699-98d999cd4f06