Fixed bug in navMacro and a bug in path. This squashes bug #46.
The problem was a comma in the return, instead of a period.
return "foo"."\n";
returns foo with a return after it
return "foo","\n";
returns just the \n probably because the comma operation comes into
play and only the result of the last operation is returned.

Apparently return doesn't grab everything to the right of it in the same
way that print does.
I'd bet that
return("foo","\n") would return a list however.

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1024 c0722133-6baf-4dd8-8699-98d999cd4f06