removed per-course logs from the initial definition of the $webworkFiles{logs} hash, and put them below, where the $courseDirs{logs} variable is defined.
(This is sort of an issue -- I want only files/dirs/URLs that are
necessarily tied to a course to be "course" files/dirs/URLs. The
transaction log (for example) could be in a location that wasn't
course-dependant just as easily. Perhaps this whole business of
"webwork" versus "course" locations is arbitrary? Should it be
rethought?)

Also fixed some formatting problems. (Mike -- I would LOVE it if you
would use spaces for aligning your equals signs, instead of tabs!)
-sam


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1032 c0722133-6baf-4dd8-8699-98d999cd4f06