Hardcopy now moves the pdf file to the tempDir/hardcopy and sends the browser a redirect to pick it up.
Cleaned up the logic of latex2pdf and generateHardcopy to
some extent.  More remains to be done.  In particular
tempDir can be generated within latex2pdf.  Possibly
it could be passed a file name as well, rather than a tex
string.

this process is a good candidate for testing the timing
mechanisms once they are done.  How much faster is
redirect?  How much faster is passing a long tex string
rather than a pointer?   etc.

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1180 c0722133-6baf-4dd8-8699-98d999cd4f06