When recording answers the answer strings are now obtained directly from the original formField entries, rather than from the processed version in the answer hashes. This insures that the memorized answers will be identical with the ones that would have been submitted from the form.
This is important for answers with multiple values which are represented by
null separated strings.  The answer evaluation process converts these strings
to references to arrays, and -- in order to have good display properties -- the
original_student_answer slot in the AnswerHash contains a representation such as
( 4, 5, 6).

This is NOT suitable for resubmitting as an answer in a form field and is therefore not the
right thing to store in the data base when saving answers.

--MIke


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1539 c0722133-6baf-4dd8-8699-98d999cd4f06