resolves bug #268:
This is a problem with the naming of the popup menu options.

"selected users" refers to the users checked in the user table, below
the action options. "users with matching user IDs:" refers to the list
of user IDs entered into the text field to the right of the popup menu.
(note the colon.)

The reporter was attempting to specify "selected users" by entering
their user IDs into the text field intended for "users with matching
user IDs".

I have changed the name of the "selected users" option to "users
selected below", which makes the distinction easier to grok.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1605 c0722133-6baf-4dd8-8699-98d999cd4f06