(1) only the content generator needed for this request is loaded, and it is done at runtime. this is good for speed and error reporting.
(2) actually instantiating CGs and running their "go" methods is defered
until the end of the handler routine. this means there is only ONE place
where this is being done.

(3) documentation has been improved.

(4) added support for UserDetail (not yet written) and
SetsAssignedToUser (written, but not yet reachable) in the UI.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1621 c0722133-6baf-4dd8-8699-98d999cd4f06