changed some variable names to make them more readable.
I've been calling variables that contain the names of courses, users,
sets, or problems things like "$courseID", "$userID", "$setID", and
"$problemID", and lists of said names "@courseIDs", "@userIDs",
"@setIDs", and "@problemIDs".

On the other hand, I call variables that contain ACTUAL RECORD OBJECTS
things like "$User", "$Set", and "$Problem", and lists of them "@Users",
"@Sets", and "@Problems". (Note the capitalization!)

This makes it totally clear whether we're talking about an OBJECT or and
ID. This is a Good Thing.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1623 c0722133-6baf-4dd8-8699-98d999cd4f06