addcourse - add a course
addcourse [--users=FILE [--professors=USERID[,USERID]...] ]
[--templates=DIR] COURSEID

Add a course to the courses directory. The required directories will be
created. Optionally, a database can be populated with users and the
F<templates> directory can be populated with the contents of another
directory. Also, one or more users can be granted professor privileges.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1653 c0722133-6baf-4dd8-8699-98d999cd4f06