I wrote a really great log message, and then I lost it. So now I'm a little pissed off.
Added code to pre-fetch users, global problems and user problems outside
of the main processing loop of scoreSet(). This allows the use of
getAllGlobalProblems() and getAllUserProblems(), which is much more
efficient than what was being done before:

For a course with 154 problems in 9 sets, all of which are assigned to
all users, scoring all sets with "IncludeIndex" and "Record Scores for
Single Sets" took the following times:

Original code: 34.07 sec
     New code:  8.68 sec

This is around four times as fast.

I initially attemped to rewrite the main loop to allow the use of
getAll*Problems(), but this proved to be time consuming and error prone.
I eventually decided to leave the main loop alone, except to replace the
DB calls inside it with accesses to pre-fetched records.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1675 c0722133-6baf-4dd8-8699-98d999cd4f06