added automatic guest logins:
With the presence of a "login_practice_user" paramter, Authen.pm will
log the user in using the first unused practice user account. Login.pm
now displays a "Guest Login" button that sends the above parameter, but
only when the course has practice users. Currently, cookies are not sent
for practice users.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1684 c0722133-6baf-4dd8-8699-98d999cd4f06