UPDATE YOUR CONFIG FILES! (global.conf.dist, database.conf.dist)
Made changes to support the storage of multiple database layouts in a
single course environment. Database layouts are now stored in a hash
named %dbLayouts in database.conf, and the default layout is aliased to
*dbLayout (see comments in database.conf.dist and global.conf.dist)
sql.conf and gdbm.conf are no longer used.

Support the selection of a specific database layout when instantiating
WeBWorK::DB. WeBWorK::DB now takes a reference to a database layout
rather than an entire course environment.

All calls to WeBWorK::DB::new were updated to pass the correct argument.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1696 c0722133-6baf-4dd8-8699-98d999cd4f06