Made modifications to PG problem editor to allow editing of course_info.
Future plans include allowing editing of login_info by the site manager (but not
by the course instructors).

All files to be edited must lie under the templates directory (for safety reasons.)

There are still issues with adequate warnings for saving files when permissions are not correctly set.

There is not yet a save as button.
--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1747 c0722133-6baf-4dd8-8699-98d999cd4f06