Corrected sorting by user_id.  Code now works correctly.  Before the sorting for this case was occuring by accident with the SQL database.
Made changes to the interface in SendMail.  The radio buttons are now horizontal to match the
radio buttons on the Instructor::Index page.  Changed Id to Login to make the sorting parameter clearer.

changed the format of the popmenu on the Instructor::Index page to match the format used on the SendMail page.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1773 c0722133-6baf-4dd8-8699-98d999cd4f06