Added utility modules for sorting lists of records and formatting records as strings (WeBWorK::Utils::SortRecords and ::FormatRecords, respectively).
These modules are important because they not only know how to sort or
format records, they can also be queried for a list of sorts or formats
suitable for display, i.e. in popup menus near the list of records.

These modules are being used as the basis for a module which
encapsulates a "record list" widget.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1823 c0722133-6baf-4dd8-8699-98d999cd4f06