Added a primitive file reading capability to Grader.pm.
Now you can attach the message from an email file
called   report_grades.msg   in the email directory which is  merged with the
file report_grades_data.csv  in the scoring directory.

there is a lot of code borrowed from SendMail and from Instructor which can be factored out.

It provides a primitive version of a feature that has been requested --
to report exam scores as well as webwork scores.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@1863 c0722133-6baf-4dd8-8699-98d999cd4f06