Added hack for allowing exporting of set definition files.
If there is a conflict the earlier files are backed up.  So set 0 is
always exported to set0.def  and a previous set0.def file moves to
set0.def.bak.    There is only one backup.

The messages need work and the utilities, particularly saveProblem should be
moved to the Utils files.

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2321 c0722133-6baf-4dd8-8699-98d999cd4f06