Moved problem-separation snippet to ABOVE the problem (to get the line above the first problem and simultaneously avoid having to check for being the last problem).
Also put the correct answers in verbatim mode to avoid having to make
and-hoc adjustments to the answer string (the result will be exactly
what the instructor typed, no matter what).

This also avoids a problem when the correct answer begins with "[",
where LaTeX would think this was part of an optional parameter to
\item and either eat the answer (up to the "]") or fail to run
entirely (if the answer were "[1,infinity)" for example).

If you don't like the verbatim, you can take it out, but then add {}
or \relax after the \item to avoid this problem.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2394 c0722133-6baf-4dd8-8699-98d999cd4f06