Provide mechanism for adding extra problem library buttons to the Library Browser.
These changes allow you to specify (via course.conf or global.conf)
that certain subdirectories of the templates directory are to have
separate buttons in the top panel of the Set Maker.  These buttons
only appear if the directories actually exist, and any other
directories are listed in the "Local Problems" button as usual.  If
the list of these directories is empty (the default), then the results
are just like they currently are, but it allows one more level of
separation of the hierarchy for those who want it.

The idea is to make symbolic links to the problem libraries from
Rochester, ASU, etc., and have separate buttons for these.  This
avoids having one monster menu with all the problems in it.  (Of
course, you could make buttons for any directories you want.)  There
are instructions in global.conf.dist that exmplain how to do configure
this.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2431 c0722133-6baf-4dd8-8699-98d999cd4f06