BIG CHANGES -- PLEASE READ.
- Changed the format of global.conf so that each setting is its own
statement. This makes it easier for users to override settings for a
course -- they can simply copy the single line from global.conf and
change the value. It also makes it easier to talk about individual
values in documentation.

- Commented each value with a description and advice. Finally! Yay!

- Moved the dvipng options to a new subhash of %pg. They are now in
$pg{displayModeOptions}{images}. Modified PG.pm and Problem.pm to match.

The next step is to do some organizaitonal work on this file so that the
most-chagned options are near the top.

Please tell me if you experience any problems as a result of these
changes! (Or better yet, submit at patch!)


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2438 c0722133-6baf-4dd8-8699-98d999cd4f06