YET ANOTHER CONFIG CHANGE -- PLEASE READ.
This chagne allows command-line scripts to continue to work without
defining a bunch of environment variables. Hopefully this will be the
last change...

Now only $webwork_dir is set in webwork.apache-config. That file then
uses this value to read the course environment and get the rest of the
configuration values that it needs out of global.conf. global.conf now
has variables like $webwork_url, $pg_dir, etc. at the top.

In summary:

* In httpd.conf, include webwork.apache-config.
* In webwork.apache-config set $webwork_dir to the path to webwork2.
* In global.conf, set $webwork_url, $pg_dir, $webwork_htdocs_url,
  $webwork_htdocs_dir, $webwork_courses_url, and $webwork_courses_dir.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2515 c0722133-6baf-4dd8-8699-98d999cd4f06