Merged changes from the following commit to rel-2-1-a1.
Revision 1.41.2.4, Tue Jul 27 22:36:57 2004 UTC (3 days ago) by gage
Branch: rel-2-1-a1
Changes since 1.41.2.3: +4 -2 lines

Allows .html, .gif, .png, .txt as well as .pg files to be uploaded and
downloaded from the template directory and files under it.  Still can't
upload/download to the html directory

This is not a final solution, but a hack that make workarounds for many
operations possible for those who do not have direct unix access.  It's
still too clunky, but it's a start.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2562 c0722133-6baf-4dd8-8699-98d999cd4f06