When a pg file has auxiliary files (like .gif images), it is often stored in a separate directory in order to keep all the files together.  In this case, the Library Browser will show the subdirectory as a separate entry in the problem directory list, even though the problem is logically part of its parent directory.
This patch modified the Library Browser so that these single-pg-file
directories are merged with their parents, as long as they also
include other non-pg-files.

It also provides a method of controling the library browser on a
directory-by-directory basis using specially named files.  If a
directory contains a file named "=library-ignore", then the directory
is never shown in the directory menu.  If it contains a file called
"=library-combine-up", then its contents are considered to be part of
the parent directory, even if it has more than one pg file.  If it
contains one called "=library-no-combine", then the directory is
always listed separately in the menu, even if it has only one pg file.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2734 c0722133-6baf-4dd8-8699-98d999cd4f06