PRELIMINARY SUPPORT FOR TIME ZONES (towards closing bug #532).
(1) new configuration variable $siteDefaults{timezone}

Set the default timezone of courses on this server. This must be a value
valid for the TZ environment variable. See also tzset(3).

If left blank, the system timezone will be used. This is usually what
you want. You might want to set this if your server is NOT in the same
timezone as your school. If just a few courses are in a different
timezone, set this in course.conf for the affected courses instead.

(2) support for second optional $display_tz argument to formatDateTime()
and parseDateTime().

If specified, formatted string datetimes will be considered to be in
that timezone instead of the server's local timezone.


Remaining work:

find all calls to *DateTime() and add the $client_tz argument.

$client_tz is is really something that *DateTime() should be able to
grab out of $ce, but they're not a method on ContentGenerator, so we
have to pass it in. It might be worth moving them to ContentGenerator.
(Or some new subclass of ContentGenerator or something?)


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2770 c0722133-6baf-4dd8-8699-98d999cd4f06