disabled "sid_open", "sid_due", and "sid_answer" formats, since they don't currenly work with timezones.
We need to find some way to get the timezone information in there
without messing up the interface. One option is to actually have the
database system aware of the timezone and propgate that information on
to the Set and UserSet Record classes. Then, add formatted pseudo-fields
to those record classes that encapsulate the date conversions. Hm, that
would be cool.

Eventually, it would be nice if the date accessors returned DateTime
objects instead of epochs, but that is for later.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2777 c0722133-6baf-4dd8-8699-98d999cd4f06