optimizations in database access (makes a big difference).
My notes:

Here's the initial timing data for a set with 10 problems:

TIMING 57300 2 1097622363.881851 (16.734902) mth162: END (assumed)

$mergedRecord is never used in FieldHTML(). However, it is populated
with a merged set or problem, which requires two database calls and a
merge operation. It appears to be OK to get rid of these calls
altogether.

TIMING 57582 0 1097624803.340321 (11.219321) mth162: END (assumed)

i'm not as worried about the several DB calls in
handle_problem_numbers(), since that only happens when "Save Changes" is
clicked. not looking at those for now...

in initialize(), the global set is fetched from the database at the
beginning (and put into $setRecord). then it is fetched again in two
places, first in a conditional if the "submit_changes" param is defined,
and then in the following conditional, when "submit_changes" is defined
and there was no error. taking this out...

(those conditionals are pretty weird to begin with. i would have
expected something more like "if (submit_changes) { ... unless (error) {
... } }". probably from grafting code together?)

ok, the big one is getting pre-fetched records into FieldHTML. I added
$setRecord and $problemRecord parameters to FieldHTML (and $Set and
$Problem parameters to FieldTable to pass through to FieldHTML), and
pre-fetch GlobalProblems, UserProblems, and MergedProblems en masse
before going through the problem IDs in body().

TIMING 58456 0 1097632191.012939 (4.516615) mth162: END (assumed)

And it's only a little more than that when viewing for a particular
user:

TIMING 58453 0 1097632169.074723 (6.513201) mth162: END (assumed)


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2913 c0722133-6baf-4dd8-8699-98d999cd4f06