Added ability to control the "Missing TeX Fonts" message generated by jsMath.  There are now two parameters in global.conf that control the message.  You can set
	$pg{displayModeOptions}{jsMath}{reportMissingFonts}

to 0 if you want no font messages, and you can set

	$pg{displayModeOptions}{jsMath}{missingFontMessage}

to a string that should be issued when the fonts are missing.  This
can be an HTML string, so you can include a link to the jsMath font
page, as the current message does.  The message will be set in a <DIV>
of class NoFontMessage, so the style attributes can be controled in
the content-generators template file.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@2983 c0722133-6baf-4dd8-8699-98d999cd4f06