These changes elevate the Options panel to ContentGenerator.pm Some of the changes are a bit hackish, but I believe that they are sufficient for the moment.  I expect much of the machinery to be replaced by an "options" table in the new database.
The current options state consists of two values:  displayMode (for
viewing equations) and showOldAnswers (for showing answers between
one session and the next).  "Sticky answers" which show form entries
when the form is submitted and returned is a feature which is always on,
as I believe it should be.

Known bugs:

The options state is lost when visiting the Instructor pages, editing
problems, and probably sending instructor feedback.  (I have made sure
that state was passed for the main links, I hope, but there are many
other possible links where state can get lost. My hope is that storing
state in the database will make it unnecessary to hunt down every link
where state needs to be preserved.)

Default options are set in the subroutine options in ContentGenerator.pm
This is called before body, where the option panel is displayed, but not
before the initialization phase.  Modules that need options in the
initialization phase (currently only Problem.pm) need to set them
themselves.

Removed File Transfer from instructor links.  File Manager works fine.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3001 c0722133-6baf-4dd8-8699-98d999cd4f06