Massive overhaul of this file to put dehackify the code, at least somewhat.
The cases for handling each file type and each action are pretty cleanly laid out.

The next reasonable step would be to reorganize the code by subclassing it.

Since in almost all cases the link knows what file type it wants to be edited
when it calls the editor, the links could specifically call the subclass editor.
This would either reduce or eliminate the need for the file_type parameter.

Properly done subclassing would make it easier to add a new file type.

Currently you need to add code in about four  places:  UI code in body,
action code in saveFileChanges,   path finding code in getFilePaths,
and redirect code in pre_header_initialize.

For a subclass this code could all be adjacent.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3037 c0722133-6baf-4dd8-8699-98d999cd4f06