Changes made to simplify the implementation of XMLRPC access to the functionality of WeBWorK.
There is still a problem with making sure that the proper warning
mechanism is passed into the problem rendering compartment.
The proper permissions for creating directories also does not
get transmitted properly.

I would like to change the way that the defineEnvirVars subroutine is
passed into the Local.pm object.  I'd like to either pass the
environment as a hash or explicitly pass a reference to define
EnvirVars rather than inherit it from PG.  In particular I'd like
to be able to define environment variables directly in RenderProblem.pm
instead of faking it by defining fake problem objects.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3073 c0722133-6baf-4dd8-8699-98d999cd4f06