Bring jsMath up to version 1.7e.
Version 1.7 introduced improved alignment in arrays, and the quality
of stretchable delimiters in some browsers.  It also fixed a couple of
bugs.  Most of the sub-versions were work-arrounds for bugs in MSIE.
Finally, there is support for controlling the style used for displayed
equations.  (Normally they are centered, but that can now be
controlled using the div.typeset CSS style.)  A number of minor
spacing problems on the PC also were corrected.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3109 c0722133-6baf-4dd8-8699-98d999cd4f06