Added styles to make Parser's ans_array methods show up properly in the student answer and correct answer blocks.
The style for the td in the results table causes havoc for tables that
are nested in there.  This solves is for Parser produced answer
arrays, but it is still messed up for the standard ans_array() and for
the output of tth (the formatted text option).

The way the style for this table is handled should probably be
changed.  Perhaps if we could override for td's in tables nested in
the results area?  I don't know enough about CSS to figure that out.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3272 c0722133-6baf-4dd8-8699-98d999cd4f06