Add command-line utility to update sql and sql_single databases to include gateway fields.  Use: wwdb_addgw [-h] [sql|sql_single]
Adds fields to the set and set_user tables in the WeBWorK mysql databases
that are required for the gateway module.  The script prompts for which
courses to modify.  Adding gateway database fields to existing courses
should have no effect on those courses, even if they are running under a
non-gateway aware version of the WeBWorK system.  (I haven't rigorously
tested this.)

If -h is supplied, the script hides the mysql password when it prompts for
it (this assumes that a Unix based stty -echo works).  If sql or sql_single
are supplied, they become the default database format.

Note that this has been tested, but not rigorously.  Please let me know
if you find that it does not perform as advertised.  It might be wise to
back up the database tables before using, and to check that it does what
we expect (that is, adds columns to the set and set_user tables) after
running it.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3388 c0722133-6baf-4dd8-8699-98d999cd4f06