I've changed errorOutput($error,$details) to accept a reference to a string or to an array for $details.  These are automatically converted to the proper form. References should be used when the $details points to a large number of bytes, such as the contents of a file.
This version is not faster than the original method where $details was always a
string, sometimes a very long string.  It does seem to save memory however.

Further savings in memory could be obtained by print directly, but there are a few instances
which call errorOutput and then do further formatting before printing.

At the moment it doesn't seem to be worth it to make the change from return to print.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3432 c0722133-6baf-4dd8-8699-98d999cd4f06