improvements to view options.
removed default options() implementation and helper function
viewOptions(), and added more modular optionsMacro() method, which is
intended to be called from content generators which require an options
panel.

POD for optionsHelper():

=item optionsMacro(options_to_show => \@options_to_show, extra_params =>
\@extra_params)

Helper macro for displaying the View Options panel.

@options_to_show lists the options to show, from among this list
"displayMode", "showOldAnswers", "showHints", "showSolutions". If no
options are given, "displayMode" is assumed.

@extraParams is dereferenced and passed to the hidden_fields() method.
Use this to preserve state from the content generator calling
optionsMacro().

This macro is intended to be called from an implementation of the
options() method. The simplest way to to this is:

 sub options { shift->optionsMacro }

=cut


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3480 c0722133-6baf-4dd8-8699-98d999cd4f06