Add the beginnings of a facility for archiving a course.  This method uses mysqldump and therefore will work only with a mysql database.  Many things are still hardwired In particular mysqldump is hardwired. The database name is hardwired to  "webwork"
Archiving consists of dumping the tables associated to the course to a file
$courseID/DATA/$courseID_mysql.database  (this file can be used
to created a new database for the course.  it will not overwrite existing
tables however.)

Then the entire course directory is tarred and gzipped and placed
in the courses directory with the naem $courseID.tar.gz

Currently nothing is deleted from the database and no directories are deleted.
So the implementation is fairly safe, but not yet very useful for course management.

There is not yet a facility for automatically importing the archived course.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3528 c0722133-6baf-4dd8-8699-98d999cd4f06