Allow snippets of html (or Tex) to be specified in global.conf to be inserted right before/after the body of a problem.  A commented out example which produces boxed problem in HTML is given.  In this approach, the added parts are not used for problems rendered in the Library Browser or Problem Set Detail.
This addresses bug 817.  This solution may not be suitable for inclusion of needed bits for modes such as asciimath or tth because the those bits of html are probably needed for display in the Library Browser.  It is hard to distinguish bits which are always wanted, and those which are only wanted for use by Problem.

To make use of this, you need to update your global.conf, and make the corresponding pg update.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3552 c0722133-6baf-4dd8-8699-98d999cd4f06