Load Parser::Legacy.
This change is needed for the current version of extraAnswerEvaluators.pl
to work out of the box.  By default, one will still get the original
(non-Parser) versions of num_cmp and fun_cmp, but turning on the
Parser-based versions will be much easier (just one line to change, or
override on a course by course basis).

There will be a companion change in the pg directories.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3571 c0722133-6baf-4dd8-8699-98d999cd4f06