Fixed a bug in which, if the Hardcopy call was the first one to need to create the  directory at (e.g.)  /ww/htdocs/tmp/myCourse then the creation of myCourse would fail.  If /ww/htdocs/tmp/myCourse/gif already existed things would go fine.
I think that more has to be done here.  We need to look at how myCourse is
created when storing a gif or html is the first action and handle these
all the same.  One of the difficulties is that the tmp directory might
be located anywhere -- it is not guaranteed to be under /ww/webwork/webwork2/htdocs
-- at least as we have defined it at the moment.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3609 c0722133-6baf-4dd8-8699-98d999cd4f06